Update patch set 16
Patch Set 16:

> Are not we actually having the issue now as ml2 sets Q_AGENT
 > itself?

We were, I rechecked the patch on top of this one that is testing these changes [0] and jobs started failing because the OVN module wasn't been sorced (calling init_ovn for example failed cause the function didn't exist).

I've added the logic now to the ml2 module and rechecked that patch and it looks good on zuul at the moment.

[0] https://review.opendev.org/#/c/735097/20

Patch-set: 16