Update patch set 8
Patch Set 8:

jobs are failing as it tries to install glance_store using pip,

To verify it is working I have added glance_store to use from git for this patch [1] and the corresponding job [2] is passing successfully.

We need to wait until new version of glance_store [3] is available which will have data files to copy.

[1] https://review.opendev.org/742373
[2] https://zuul.opendev.org/t/openstack/build/24d3dca407eb4fee82ed52f02e833c10
[3] https://review.opendev.org/745796

Patch-set: 8