cinder: add lsscsi as dependency
This is an os-brick dependency listed in its bindep.txt file.
os-brick is usually installed from pip, so its bindep.txt file
is not available.
While thinking about a more robust solution to prevent the
keeping track of the dependencies in two places (bindep and here),
(temporarily?) duplicate them, as it has been the case for
the other os-brick dependencies already listed here.

Depends-On: https://review.opendev.org/745752
Change-Id: I9ab6e215dbef9ebdb1946da2f9a40ce020ecc95b