Set devstack-platform-centos-8 voting
It is important for projects to do CI/CD gating on the top most common
environment used by operators.

CentOS 8 is one of three LTS or stable distributions for Victoria [2]
and the most widely used distribution according to the OpenStack User
Survey 2019 [1] with a combined 55% of OpenStack cloud deployments
running mainly on CentOS/RHEL distribution. We should gate on CentOS 8.

Should this platform become a gate blocker, the community should either
fix the issue internally when applicable, or report it upstream and
provide a workaround whenever possible. If a workaround is not feasible
and upstream is not responsive, please escalate to CentOS/RHEL
colleagues in the OpenStack community; only after the job could be
turned non-voting (and back to voting once issue is resolved).

[1] https://www.openstack.org/analytics
[2] https://governance.openstack.org/tc/reference/runtimes/victoria.html

Change-Id: I92325d11998f6a130266fdaaa2d89f460e55b7c9