Update patch set 2
Patch Set 2:

> "Having interest in" isn't the same as actually showing
 > contributions to day to day operations. The centos8 job took ages
 > to be brought into working shape, that doesn't make me very
 > confident in its future maintenance.

"Having interest in" was a polite way to say those folks are expected to contribute to smooth devstack deployments on CentOS just like they *do* contribute today for Ubuntu-based deployments.

 > For starters you could take a
 > look at this https://review.opendev.org/746590.

That is a backport patch and it LGTM (+ a rebase on top of a backport of this patch eventually).

Patch-set: 2