Update patch set 2
Patch Set 2:

> i think this makes sense but i think we ahve also already move to
 > focal so the ci jobs wont test this except grenade on master right?

Well I'd like to bump things before we move to focal to avoid a last minute rush around M3.

 > the greade job still uses bionic
 > https://zuul.opendev.org/t/openstack/build/6f807b8d64ca435190c0b39975e9d31c/log/job-output.txt#47
 > 
 > but we seam to have somehow skipped your change
 > 
 > https://zuul.opendev.org/t/openstack/build/6f807b8d64ca435190c0b39975e9d31c/log/job-output.txt#5138-5158
 >
 > i think your chagne looks correct but we shoudl figure out why
 > grenade is not using it.

That's because the initial grenade run is using the stable/ussuri branch without this change, it's working on the master branch. I'm more concerned with the errors in tempest-full-py3.

Patch-set: 2
Label: Verified=0