Add etcd3gw to managed libs
Now that etcd3gw is hosted on OpenDev [1], let's make it
a first-class lib in DevStack too and allow source testing.

etcd3gw is the default tooz driver for coordination in Cinder.
It uses etcd3 gateway API.

This patch includes extra comment fixes.
We could probably drop the installation of etcd3 as it is not
a usable candidate anyway.

[1] https://review.opendev.org/747185

Change-Id: Iebc263172666dcdadc41c34832b4810b3e60a7ba