Allow extra servicefile options in run_process
This commit makes it possible to allow setting additional options or
overwriting defaults in servicefiles created by run_process. This is
useful e.g. for kuryr-kubernetes gate that's running Kubernetes services
which can crash due to etcd issues (those are really common on the gate
setups due to storage latency). With this change it should be possible
to set `Restart=Always` to make sure systemd is able to recover from the
issue.

Change-Id: Iabeb7a697e60a97820ed56c816a41306ab503655