Update patch set 3
Patch Set 3:

I've never pushed on this because I've never really felt like I could commit to fixing anything that might happen to the build in a timely fashion that would unblock the gate.  And then it has a history of "we tried this voting, but it didn't work", so it's harder to get anyone to believe next time it would ever be any different.

Especially because this has a big dependency on the RDO/cloud-sig packages -- we necessarily trail what happens there (look how long centos 8 took).

If I can add any piece of concrete information that would be

https://people.redhat.com/~iwienand/devstack-status/

which tracks the centos-8 runs.  It's pretty green.

If other people are saying they will try to keep it green, well that's probably already more than I've heard before.  So I'm happy to try it, but open to discussions if this should be after release, etc.

Patch-set: 3
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Review-Priority=0