Update patch set 1
Patch Set 1: Workflow-1

This is happening on a CentOS 8 job [1]
The error appears when migrating a VM from controller 1 to controller 2. The fix seam it is not working. The problem is still present after this fix. The issue is not present on Ubuntu Bionic or Focal. I suspect the problem is therefore related with SELinux or DevStack.



[1] https://zuul.opendev.org/t/openstack/build/f34857e95a3542d2803119a2d2ad1556
[2] https://storage.gra.cloud.ovh.net/v1/AUTH_dcaab5e32b234d56b626f72581e3644c/zuul_opendev_logs_f34/740929/67/check/tobiko-devstack-faults-centos-8/f34857e/compute2/logs/screen-n-cpu.txt

Patch-set: 1
Label: Workflow=-1