Update patch set 1
Patch Set 1:

Let me check first the exact failure and see how it can be addressed. No job name is mentioned in any commit message of the relevant reviews (this one and the other fix attempt), only a comment mentions openstacksdk-functional-devstack-ironic.
I'd also like to note that breaking other voting jobs is not a solution either. We can all agree on that, I hope. Moreover, and that the proper way would be to file a bug first.

Patch-set: 1