Update patch set 1
Patch Set 1:

Looks OK to me and might reduce the rate at which the failure occurs but just FYI because of a possible regression (in provider nodes) failures will still be seen with this patch. TL;DR is that presence of a IPv4 default route is supposed to be safe to assume and has been until recently, and it is unknown why not yet.

Another patch/fix has been proposed to additionally look for a IPv6 default route to handle remaining failures:

https://review.opendev.org/761178

but there is discussion in the review about whether it's correct to fix it in this way (see review for details).

And here's some more discussion on IRC about the problem and how something must be wrong with limestone nodes:

http://eavesdrop.openstack.org/irclogs/%23opendev/%23opendev.2020-11-05.log.html#t2020-11-05T22:37:00

Given that there's apparently something wrong with the limestone provider nodes and this patch will not fix all cases and might not be the recommended approach by the infra team, I'm conflicted about whether it should be backported right now.

Patch-set: 1
Reviewer: Gerrit User 4690 <4690@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Review-Priority=0