Update patch set 2
Patch Set 2:

we need to upgrade the libvirt version in ussuri or we will have the intermitent failure due to libvirt <6.0.0 using the legacy -drive option.

the already merged patches will unplug my changes on master but we are going to hit the same issue when we backport.

i dont see a reason to test with the previous uca i dont think we have every had that policy we have to my knowladge always enabled the version we needed to provide the required depencies or workaround known bugs.

Patch-set: 2