Update patch set 2
Patch Set 2:

the minium version for ussurig is below what we are currently testing
its libvirt 4.0.0

https://github.com/openstack/nova/blob/stable/ussuri/nova/virt/libvirt/driver.py#L225-L226

we are currently testing with 5.2 due to other issues.
https://review.opendev.org/#/c/752641/

the only reason that patch did not bump to ussui at that point was we believed

https://bugs.launchpad.net/nova/+bug/1882521 was caused by the new qemu version on focal but in fact it was not which is explained in 
https://review.opendev.org/#/c/757306/ it was actually a race we had and undocumented qemu behaviour.

if you are deploying on ubuntu in production and installing form packages you are going to be enabling the ubuntu cloud archive for that release to get the packages for that release.

so the ussuri code would normally be deployed using ussuri generally as a minimum. that said the code would work with older versions and newer versions but that would imply you are installing form pip not the disto packages.

nova cannot fix the libvirt issue since we cannot control how the libvirt xml we generate is translated into qemu args.

so we would have to disable all paused live migration testing on branch that dont have 6.0.0 or mark the test in some what to not count towards the result e.g. as flaky to ignore failures. currently tempest cant do that i believe.

anyway for now ill proceed with the nova bug fix and hopefully the grenatde job on master will pass but we likely will need to do something when we backport.

Patch-set: 2