Do not add links to upper-constraints.txt but filter packages
During installation upper-constraints.txt is edited: package names that
need to be installed from local source, are replaced with 'local link'.
With this editing we avoid to accidentally install a package from pypi,
that should be installed from local source, if it is pulled in as
dependency of any package.

To be able to use links in constraint files was an undocumented and
unintentional feature [1], which behavior was removed with the newly
introduced resolver of pip, in version 20.3 [2].

Since the packages that are available in local source are installed
anyway, we can simply filter out these package names from the
constraint file.

[1] https://github.com/pypa/pip/issues/8210
[2] https://pypi.org/project/pip/20.3/

Closes-Bug: #1906322
Change-Id: I6929b0f971ea72e5d7b77d85219db069431c48d3