Update patch set 10
Patch Set 10: Code-Review-1

(1 comment)

I tried that patch on Ubuntu 20.04 but devstack is failing for me with error like:

```
+lib/neutron_plugins/ovn_agent:_start_process:190  sudo systemctl enable ovn-northd.service
The unit files have no installation config (WantedBy=, RequiredBy=, Also=,
Alias= settings in the [Install] section, and DefaultInstance= for template
units). This means they are not meant to be enabled using systemctl.

Possible reasons for having this kind of units are:
• A unit may be statically enabled by being symlinked from another unit's
  .wants/ or .requires/ directory.
• A unit's purpose may be to act as a helper for some other unit which has
  a requirement dependency on it.
• A unit may be started when needed via activation (socket, path, timer,
  D-Bus, udev, scripted systemctl call, ...).
• In case of template units, the unit is meant to be enabled with some
  instance name specified.
```

So I think that this may be related to that change :)

Patch-set: 10
Reviewer: Gerrit User 11975 <11975@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1