Update patch set 10
Patch Set 10:

> Patch Set 10: Code-Review-1
> 
> (1 comment)
> 
> I tried that patch on Ubuntu 20.04 but devstack is failing for me with error like:
> 
> ```
> +lib/neutron_plugins/ovn_agent:_start_process:190  sudo systemctl enable ovn-northd.service
> The unit files have no installation config (WantedBy=, RequiredBy=, Also=,
> Alias= settings in the [Install] section, and DefaultInstance= for template
> units). This means they are not meant to be enabled using systemctl.
> 
> Possible reasons for having this kind of units are:
> • A unit may be statically enabled by being symlinked from another unit's
>   .wants/ or .requires/ directory.
> • A unit's purpose may be to act as a helper for some other unit which has
>   a requirement dependency on it.
> • A unit may be started when needed via activation (socket, path, timer,
>   D-Bus, udev, scripted systemctl call, ...).
> • In case of template units, the unit is meant to be enabled with some
>   instance name specified.
> ```
> 
> So I think that this may be related to that change :)

Thanks Slaweq, 

I was able to reproduce the issue after a few tries, I believe the problem was that in Ubuntu there's a "ovn-central.service" service file which is used to start both northd and the OVN DBs in the right order. So I changed the code on Ubuntu to use it. 

Note that on CentOS the ovn-northd.service is the service file that does it, there's no ovn-central.service there.

Patch-set: 10