disable apic emulation for cirros images with libvirt.
This change modifies the default image url for libvirt
to use the uae cirros image. the uae image have a
blank root disk with a seperate initrd and kernel.
This allow nova to use libvirts direct kernel boot
feautre which enables the use fo the os_console_command
image property to disable the guest apic via the kernel
commandline.

This works around a know ubuntu kernel bug that is unpatched
in the 5.3 kernel used by cirros resulting in periodic
kernel hangs in the ci.

Change-Id: I2f4cd8e670f5a78fa8c72368ea47959bdcd28b94