Update patch set 33
Patch Set 33:

The latest patch-set should work with both CentOS and Fedora. The problem I found was just that the pre-set configuration that comes with the packaging was conflicting with the configuration we were doing in the DevStack module.

Patch-set: 33
Reviewer: Gerrit User 6773 <6773@4a232e18-c5a9-48ee-94c0-e04e7cca6543>