WIP: Configure Cinder backup driver
This patch adds a new environment variable, CINDER_BACKUP_DRIVER for
configuring cinder backup driver. This gets cinder backup driver
configurable with a similar pattern to cinder backends.

The following backup drivers can be configured:
swift:
This is the default backup driver.
ceph:
This already can be configured if ceph backend driver is enabled. For
backward compatibility, ceph backup driver is used if ceph backend
driver is enabled and no backup driver is specified.
s3:
This backup driver gets configurable with this patch.

Change-Id: I356c224d938e1aa59c8589387a03682b3ec6e23d