WIP: Async task support
I know this is going to be controversial, but hear me out.

We have a *ton* of stuff in devstack that is very linear, specifically
the ten-ish minutes we spend loading osc to run a single API command
against something. We also generate configs, sync databases, and other
things that use one core of our worker and make our runtime longer
than it really needs to be.

The idea in this patch is to make it super simple to run some things
in the background and then wait for them to finish before proceeding
to something that will require them to be done. This does (of course)
introduce some interleaving of the tracing output. However, in the
process of writing and debugging this I found it surprisingly not
very difficult to work out what failed earlier than the actual exit
trap. Some extra logging around start/stop of the tasks makes it
pretty clear and easy to search for. The tracing tends to be quite
chunky so you're not getting every-other-line interleaving.

If this has legs, here are some things I'd plan to do:

- Make it disabled by default (at least for now). When disabled,
  async_run would just be a synchronous call straight through,
  so no difference from today. async_wait() would be a nop.
- Make it enable-able, perhaps by individual async future name
  so that we can turn individual things on and off.
- Investigate logging the activity of an async job separately, either
  for later collection, or to merge into the main output in one
  chunk.
- Report uncollected futures at the end of the stack
- Provide some way to make sure we don't start more than $ncpus
  tasks at once (if desirable).
- Look for other places we could refactor serialized work for
  better parallelism.

On my system, this results in a much higher load average while running
devstack (about 2.3 vs 0.5 normally), indicating more is getting
done. On my older rotating rust devstack machine with 8 cores, this
runs 13% faster on a stripped-down nova-glance-neutron-keystone
config. Results in the gate will, of course, dictate the usefulness,
but with minimal refactoring for parallelism, 13% speedup seems worth
consideration to me.

Change-Id: I270a910b531641b023c13f75dfedca057a1f1031