WIP: Async task support
I know this is going to be controversial, but hear me out.

We have a *ton* of stuff in devstack that is very linear, specifically
the ten-ish minutes we spend loading osc to run a single API command
against something. We also generate configs, sync databases, and other
things that use one core of our worker and make our runtime longer
than it really needs to be.

The idea in this patch is to make it super simple to run some things
in the background and then wait for them to finish before proceeding
to something that will require them to be done. This avoids the
interleaving you would expect by redirecting the async tasks to a log
file, and then cat'ing that log file synchronously during the wait
operation. The per-task log file remains so it's easier to examine
it in isolation.

If this has legs, here are some things I'd plan to do:

- Make it disabled by default (at least for now). When disabled,
  async_run would just be a synchronous call straight through,
  so no difference from today. async_wait() would be a nop.
- Make it enable-able, perhaps by individual async future name
  so that we can turn individual things on and off.
- Provide some way to make sure we don't start more than $ncpus
  tasks at once (if desirable).
- Look for other places we could refactor serialized work for
  better parallelism.

On my system, this results in a much higher load average while running
devstack (about 2.3 vs 0.5 normally), indicating more is getting
done. On my older rotating rust devstack machine with 8 cores, this
runs 13% faster on a stripped-down nova-glance-neutron-keystone
config, and about 16% on a fullish config. Results in the gate will,
of course, dictate the usefulness, but with minimal refactoring for
parallelism, ~15% speedup seems worth consideration to me.

Change-Id: I270a910b531641b023c13f75dfedca057a1f1031