Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> I'd rather go forward with https://review.opendev.org/c/openstack/devstack/+/765149 directly. Can you verify that that patch also solves the tempest issue?

It doesn't for me:

++lib/tempest:install_tempest:709           /opt/stack/tempest/.tox/tempest/bin/pip install -c /opt/stack/requirements/upper-constraints.txt -r requirements.txt
DEPRECATION: Constraints are only allowed to take the form of a package name and a version specifier. Other forms were originally permitted as an accident of the implementation, but were undocumented. The new implementation of the resolver no longer supports these forms. A possible replacement is replacing the constraint with a requirement.. You can find discussion regarding this at https://github.com/pypa/pip/issues/8210.
ERROR: Links are not allowed as constraints

$ git rev-parse HEAD
5e408844a402116b4f500dd6b8c271d258381e4f

It's frustrating to be working on devstack and have to keep patching this fix underneath on local trees. Can we just land this and undo it when there's a proper fix?

Patch-set: 1