Update patch set 1
Patch Set 1:

The reason I'm not easily applying this patch is that we had a similar situation when we capped pip<10 and it took years to get rid of that again.

So my next question would be why we don't see this issue in gate tests. Can we add a test or job that covers your scenario? Otherwise we'd have no way of avoiding regressing again once we have what looks like a proper fix, like the one I already assumed we had with Elod's patch.

Patch-set: 1
Reviewer: Gerrit User 13252 <13252@4a232e18-c5a9-48ee-94c0-e04e7cca6543>