Update patch set 1
Patch Set 1: Review-Priority+2

I know this is not perfect fix and it might take time to have actual fix like frickler mentioned or pip<10 case. And for tempest venv it would not be just one place fix as 2-3 place in devstack and greande and tempest/devstack plugins side, that venv is being run with file as constraint.

So until we fix all those places which can take time I agree to merge this and unblock the local run. and I am sure it can start happening in gate at any time depends on virtualenv version of way of installing venv. raising priority for this to decide on it fast.

Patch-set: 1
Reviewer: Gerrit User 8556 <8556@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Review-Priority=+2