[OVN] Remove old Q_BUILD_OVS_FROM_GIT flag
The Q_BUILD_OVS_FROM_GIT flag is a leftover from when OVN in the same
repository as OVS (before the split) and does not make sense anymore.

The flag was being used to skip the install_ovn() function entirely
but, now it makes no sense anymore because this method will install OVN
by either compiling it from source (if OVN_BUILD_FROM_SOURCE is True)
or packages and skipping it means no OVN will be installed (since it no
longer comes packaged together with OVS).

Change-Id: I92ebef2634cc1f4ee9f266ba9e34c17a4b8b90af
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>