Fix up some of the error handling after wait
When I changed the logging stuff and made the tracing non-interleaving,
I didn't quite get the failure case correct to make a failure easy
to debug. This fixes that by making sure we don't errexit before
we get a chance to show what failed.

Also, grab the async log directory in zuul for forensics.

Change-Id: I0b0787e56b4f16f22bb8035cf4c71855ca0e6f11