Increase volumes quota for service project
If we are backing glance with cinder, we will use more volumes and
if timing is right, we will clash with other tests and be unable
to create what we need. If we are backing glance with cinder, we
should increase the volumes quota, which this patch does (to 50 from
a default of 10).

Change-Id: I2ad1c4d21f996ee1a9ce29ba4f1a4b8f5720f8fb
Closes-Bug: #1914665