Update patch set 20
Patch Set 20:

Okay, sorry for disappearing on this. The recent mailing list thread about gate resources prompted me to revisit it.

I reimplemented this as a service in devstack instead of an Ansible wrapper around it. I think that cleans things up significantly and makes it simple to avoid using this when openstackclient patches are in the dependency tree. I've temporarily enabled it in .zuul.yaml just for demonstration purposes, but I think that should be removed before it merges so everyone doesn't inherit it automatically. If we can turn it on for some of the projects that run the most jobs that should get us a lot of benefit without having to deal with lots of corner cases.

Patch-set: 20
Reviewer: Gerrit User 6928 <6928@4a232e18-c5a9-48ee-94c0-e04e7cca6543>