WIP: Ensure cinder has an endpoint for system-users
All the endpoint URLs for cinder have project ID templating. This
prevents the endpoints from making it into service catalogs for
system-scoped tokens, or system users.

This commit attempts to work around that issue by creating a cinder
endpoint with the project ID templating removed, allowing it to show up
in service catalogs for system users.

Change-Id: Ia2b6706d3270066b5fc11817d0cba76e18551eff