Update patch set 1
Patch Set 1: Code-Review+1

Thanks Dan for fixing this! The patch is failing in the grenade jobs, as Stein needs this patch as well (and needs to be merged in reverse order). I've cherry picked this fix to stable/stein. Grenade jobs are set as non-voting in Stein, so hopefully that can be merged quickly.

Patch-set: 1
Reviewer: Gerrit User 17685 <17685@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1