Update patch set 11
Patch Set 11:

(2 comments)

> I feel like the start/stop is a bit fragile and we have mechanisms to write out the service file, which can then be started separately; see inline.

Understand, and I should have circled back to try to clean that stuff up or figure out how to make it do what I want. There was a lot of drama in landing the glance part of this feature, which made it very late and I avoided poking the QA people to look at this urgently while that was going on. Unfortunately, we've now got the feature landed and really want to get it exposed to the "firehose" before the release, but we can't without this.

I'll circle back to this tomorrow and see if I can reproduce the original problem with iniset, if you don't first.

If there's any chance we can just land this as-is so that we can have this tested in the meantime (localized to the glance tree), I'll surely commit to following up. I have a pretty good record on that.

Patch-set: 11