cinder: Default CINDER_ISCSI_HELPER to lioadm
As outlined in bug #1917750 the use of tgtadm in multinode environments
with multiple c-vol services can cause volumes to use duplicate WWNs.

This has been shown to cause some encrypted volume test failures as
os-brick returns a /dev/disk/by-id path to n-cpu that can point to the
wrong underlying volume when multiple volumes with the same WWN are
connected to a host.

There is also some speculation that the duplicate WWNs are also causing
libvirt to fail to detach volumes from instances but as yet this has not
been proven.

This change aims to avoid all of the above by switching the default of
CINDER_ISCSI_HELPER to lioadm for all deployments instead of just EL and
SUSE based deployments.

The Bionic platform job however is pinned to tgtadm as there issues
installing python3-rtslib-fb.

The issue is cherry-picked to wallaby and before due to #2081177
which hitting after recent kernel updates in ubuntu focal.

Closes-Bug: #1917750
Related-Bug: #2081177
Change-Id: If5c860d1e69aaef9a9236303c370479a7714ad43
(cherry picked from commit c062792709def9ef10ddac68867e1b7bf9009435)
Conflicts: .zuul.yaml