cinder: Default CINDER_ISCSI_HELPER to lioadm on Ubuntu
As outlined in bug #1917750 the use of tgtadm in multinode environments
with multiple c-vol services can cause volumes to use duplicate WWNs.

This has been shown to cause some encrypted volume test failures as
os-brick returns a /dev/disk/by-id path to n-cpu that can point to the
wrong underlying volume when multiple volumes with the same WWN are
connected to a host.

There is also some speculation that the duplicate WWNs are also causing
libvirt to fail to detach volumes from instances but as yet this has not
been proven.

This change aims to avoid all of the above by switching all Ubuntu based
deployments over to lioadm by default.

The check ensuring lioadm is used on EL and SUSE based deployments is
extracted to allow Ubuntu based deployments to use tgtadm for testing.
Given the above this should only ever happen in single node
environments.

The Bionic platform job has been pinned to tgtadm as there issues
installing python3-rtslib-fb.

Closes-Bug: #1917750
Change-Id: If5c860d1e69aaef9a9236303c370479a7714ad43