Use (or set properly) system-id generated by openvswitch
In case when OVN_UUID isn't set by used, and it isn't stored
in ovn-uuid file, Devstack will first check if it is already
generated and stored in /etc/openvswitch/system-id.conf file.
If it's not, it will generate it.
Finally this OVN_UUID value will now be storer in the
/etc/openvswitch/system-id.conf file so it can be set to same value
after openvswitch will be e.g. restarted.

Closes-Bug: #1918656
Change-Id: I8e3b05f3ab83e204bc1ce895baec0e1ba515895b